Leftover Empty Folder Patch
Empty folders are now deleted as well after removal of all files yay.

+++ Archi_Maid.py
@@ -14,15 +14,27 @@ with zipfile.ZipFile(zippath, 'r') as zip_ref:
         print(file)
         main_path=os.path.join(extract_loc, file)
 
-        if os.path.exists(main_path):
-            if os.path.isdir(main_path):
-                shutil.rmtree(main_path)
-                print(f"Removed Folder: {file}")
-            else:
-                os.remove(main_path)
-                print(f"Removed: {file}")
+        if os.path.exists(main_path) and os.path.isfile(main_path):
+            os.remove(main_path)
+            print(f"Removed: {file}")
         else:
             print(f"File not found (T_T) : {file}")
+            
+    folders=set()
+    for file in file_list:
+        chunks=file.replace("\\", "/").split("/")   
+        for i in range(len(chunks)):
+            parentfolder="/".join(chunks[:i])
+            if parentfolder:
+                folders.add(os.path.join(extract_loc, parentfolder))
+    
+    for file in sorted(folders, reverse=True):
+        if os.path.exists(file) and os.path.isdir(file):
+            try:
+                os.rmdir(file)
+                print(f"Removed Folder: {file}")
+            except OSError:
+                print(f"Skipped (folder has stuff inside) : {file}")
 
 '''
     for file in file_list:
